cmd/go: fix toolchain semantics to match design doc
The design doc says 'toolchain' lines apply even if the default
toolchain is older than the one specified in the toolchain line.
Respect that again. (For a while I thought it was going to lead to
an unfortunate non-commutativity if 'go get' added toolchain
lines, but that turned out not to be the case.)

Also clean up the tests quite a bit.

Finally drop + from the acceptable version suffixes; we use + for +auto and +path.

For #57001.

Change-Id: Ia92c66be75d6d0e31cb4e2c0aa936fa4ec5c0a8c