go/types, types2: don't infer type argument for unused parameter in interfaces
Two interface types that are assignable don't have to be identical;
specifically, if they are defined types, they can be different
defined types. If those defined types specify type parameters which
are never used, do not infer a type argument based on the instantiation
of the matching defined type.

Fixes #60377.

Change-Id: I91fb207235424b3cbc42b5fd93eee619e7541cb7