cmd/compile: print error ULRs together with error messages
DO NOT SUBMIT.

Enable the printing of error URLs of the form

        [go.dev/e/ddd]

where ddd is a (decimal) error code. The link appears at the end
of (the first line of) an error message reported by the compiler.

The link refers to documentation explaining the error in more detail.

Printing of the URL can be disabled by setting the compiler flag

        -errorurl=false

Change-Id: I50c6a365fa7c3ae8208cf62ef73bd60697179fb6