io: ReadFile: don't check for slice re-allocation in the first iteration
At the beginning of the for-loop iteration size > 0 always. so in
the first iteration, this check is not necessary. we can move this check to
after the read operation.

besides this change >= operator also can be simplified to == in the if-statement

Change-Id: I205e4a842ced74f31124b45a39b70523b56ad840
GitHub-Last-Rev: 1c568d229551eca3a7b6fbb51b76ec6cc7e8e899
GitHub-Pull-Request: golang/go#60473