os: ReadFile: don't check for re-allocation in the first iteration
At the beginning of the for-loop iteration size > 0 always. so in
the first iteration, this check is not necessary. we can move this check to
after the read operation.

besides this change >= operator also can be simplified to == in the if-statement

Change-Id: I205e4a842ced74f31124b45a39b70523b56ad840
GitHub-Last-Rev: 2fdf25dff2e9984d3a8f8e5e612ea802c88e88a1
GitHub-Pull-Request: golang/go#60473