cmd/compile: fix location lists that extend over stack growth code
This patch fixes some incorrect DWARF location lists. Before this patch,
a location list for a variable that was available until the "end of its
function" erroneously extended over trailing code that dealt with the
stack growth code. The code in question is placed at the end of the
function, but it is morally part of the function's prologue (the
prologue jumps to it if stack growth is needed, and this trailer then
jumps back to the prologue). Thus, for the purposes of the location
lists, it should be treated more like the beginning of the function than
the end of the function; instructions for reading variables at the end
of the function do not apply in this code region.

The patch makes the loclists for regular variable not expand over this
trailing code, as they do not yet exist when this code runs. Function
arguments are special though - they do exist when the stack growth code
runs, and generally can be read. For function arguments, this patch
"fixes up" the location lists generated by the compiler after the linker
has inserted the stack growth code, as described below:

The layout of a function's code is as follows:

<function prologue jumps to the stack growth trailer>
<regular function code>
...
ret
movq    %rax, 0x38(%rsp)                  <- FuncLogicalEnd
movq    %rbx, 0x40(%rsp)
...
callq   <runtime.morestack_noctxt.abi0>   <- StackGrowthCall
movq    0x38(%rsp), %rax
movq    0x40(%rsp), %rbx
...
jmp     <beginning of function>
int3                                      <- FuncPhysicalEndA

The stack growth trailer consists of register spilling (starting at
FuncLogicalEnd), a call to the runtime stack growth function (marked by
StackGrowthCall), and then register unspilling (up until
FuncPhysicalEnd).
A variable that's available at the beginning of the function (i.e. a function
argument) is available in between [FuncLogicalEnd,StackGrowthCall) using the
same DWARF instructions as the ones used at the beginning of the function.
The variable is then available in between [StackGrowthCall,FuncPhysicalEnd)
with new instructions referencing the stack location where the variable was
spilled to.

The patch only performs this fixup for function arguments on amd64 and
arm64; the tracking that linker needs to do for identifying
FuncLoficalEnd and StackGrowthCall is architecture-specific and I didn't
go through the trouble of doing it outside these two archs.

Fixes #60493

Change-Id: I8ccc9ae802bddc9427ecc4ffc5a26f172a8d6b83