net: fix panic when calling net.Listen or net.Dial on wasip1
Address a panic that was caused by net.Dial/net.Listen entering the fake
network stack and assuming that the addresses would be of type *TCPAddr,
where in fact they could have been *UDPAddr or *UnixAddr as well.

Three new tests running exclusively when compiling with GOOS=wasip1 are
added as well to verify that code calling these functions of the net
package receive an error indicating that the platform does not support
creating sockets rather than panicking.

Fixes #60739

Change-Id: I2688f1a0a7c6c9894ad3d137a5d311192c77a9b2