testing: improve benchmarking example
In #27400 the concern is that the current documentation and
examples in testing don't document best practices to avoid
optimizations in benchmarks. This CL changes the example to
explicitly mention a possible option to avoid it realistically.

I chose to use the sink approach vs. runtime.KeepAlive because
the former is used in many stdlib benchmarks, while I didn't
find much examples of the latter. It's also slightly easier
to use because each b.N iteration can write into the sink and
we don't have to worry about collecting all the results of
all the iteration to "keep alive".

Change-Id: I3cfdde0ce12739a671386433679dadcb0e42dc5f