cmd/go: reset CC in build_darwin_cc_arch test
The test currently clears CC before running, in case user has it
set. But if I do "CC=mycc ./all.bash", mycc is baked into the
toolchain as DefaultCC, and clearing CC at the test will make it
use DefaultCC. As the test expects clang, just set it to clang.

Change-Id: I7e2148f354c7b9befcf63356847aab46d3d05b1e