[dev.inline] cmd/compile/internal/inline: add framework to compute function "properties"
Add some machinery to support computing function "properties" for use
in driving inlining heuristics, and a unit testing framework to check
to see if the property computations are correct for a given set of
canned Go source files. This CL is mainly the analysis skeleton and a
testing framework; the code to compute the actual props will arrive in
a later patch.

Change-Id: I7970b64f713d17d7fdd7e8e9ccc7d9b0490571bf