[dev.inline] cmd/compile/internal/inline: panic/exit flags analysis for inline heuristics
Add code to compute whether a given function appears to
unconditionally call panic or exit (as a means of driving inlining
decisions). Note that this determination is based on
heuristics/guesses, as opposed to strict safety analysis; in some
cases we may miss a function that does indeed always panic, or mark a
function as always invoking panic when it doesn't; the intent is get
the right answer in "most" cases.

Change-Id: Ibba3e60c06c2e54cf29b3ffa0f816518aaacb9a3